Moved directories into /src
