+++ server/src/index.ts
@@ -5,10 +5,10 @@ import morgan from 'morgan';
 import { port } from './config';
 
 import { ExpressError, NotFoundError } from './errors';
-import { router as loginRoute } from '../routes/login';
-import { router as profileRoute } from '../routes/profile';
-import { router as stockRoute } from '../routes/stock';
-import { authenticateJWT } from '../middleware/auth';
+import { router as loginRoute } from './routes/login';
+import { router as profileRoute } from './routes/profile';
+import { router as stockRoute } from './routes/stock';
+import { authenticateJWT } from './middleware/auth';
 
 const app = express();
