+++ server/src/routes/login.ts
@@ -2,7 +2,7 @@ import { Router } from 'express'
 
 import { Discord } from '../api/discord';
 import { createToken } from '../helpers/token';
-import { BadRequestError } from '../src/errors';
+import { BadRequestError } from '../errors';
 
 export const router = Router();
